modificações nas tabelas

+++ database/migrations/2019_05_23_184651_create_locations_table.php
@@ -15,8 +15,8 @@ class CreateLocationsTable extends Migration
     {
         Schema::create('locations', function (Blueprint $table) {
             $table->bigIncrements('id');            
-            $table->double('lat');
-            $table->double('lng');
+            $table->float('lat');
+            $table->float('lng');
             $table->timestamps();
         });
     }
